requirements.txt
pyrogram==2.0.106
pymongo[srv]==4.5.0
python-dotenv==1.0.0
aiohttp==3.8.5
